+++ natsort/__init__.py
@@ -7,7 +7,8 @@ from natsort.natsort import (natsort_key, natsort_keygen, ns,
                              natsorted, humansorted, versorted,
                              realsorted, index_realsorted,
                              index_natsorted, index_versorted,
-                             index_humansorted, order_by_index)
+                             index_humansorted, order_by_index,
+                             decoder, as_ascii, as_utf8)
 from natsort._version import __version__
 
 __all__ = [
@@ -22,5 +23,8 @@ __all__ = [
     'index_humansorted',
     'index_realsorted',
     'order_by_index',
+    'decoder',
+    'as_ascii',
+    'as_utf8',
     'ns',
 ]