+++ docs/source/changelog.rst
@@ -3,7 +3,15 @@
 Changelog
 ---------
 
-04-04-2015 v. 3.5.4
+04-06-2015 v. 3.5.6
+'''''''''''''''''''
+
+    - Added 'UNGROUPLETTERS' algorithm to get the case-grouping behavior of
+      an ordinal sort when using 'LOCALE'.
+    - Added convenience functions 'decoder', 'as_ascii', and 'as_utf8' for
+      dealing with bytes types.
+
+04-04-2015 v. 3.5.5
 '''''''''''''''''''
 
     - Added 'realsorted' and 'index_realsorted' functions for